requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
attr==0.2.0
certifi==2017.4.17        # via requests
chardet==3.0.4            # via requests
click==6.7
idna==2.5                 # via requests
pyyaml==3.12
requests==2.17.3
urllib3==1.21.1           # via requests
